requirements.txt
line-bot-sdk==1.15.0
Fkask==1.1.1
pya3rt==1.1
